#72 refactor: getAuthentication 로직을 UserService로 통합
