+++ src/main/java/welkit_server/domain/admin/service/CommunityManagementService.java
@@ -7,15 +7,12 @@ import org.springframework.transaction.annotation.Transactional;
 import welkit_server.domain.admin.dto.response.*;
 import welkit_server.domain.community.entity.CommunityPosts;
 import welkit_server.domain.community.repository.CommunityPostRepository;
-import welkit_server.domain.user.entity.User;
-import welkit_server.domain.user.repository.UserRepository;
+import welkit_server.domain.user.service.UserService;
 import welkit_server.global.exception.message.ErrorMessage;
 import welkit_server.global.exception.model.NotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.PageRequest;
-import welkit_server.global.exception.model.UnauthorizedException;
-import welkit_server.global.security.dto.CustomUserDetails;
 import java.util.List;
 
 @Service
@@ -24,10 +21,10 @@ import java.util.List;
 public class CommunityManagementService {
 
     private final CommunityPostRepository communityPostRepository;
-    private final UserRepository userRepository;
+    private final UserService userService;
 
     public CommunityManagementPostResponse getAllSanctionPosts(int page, int size, Authentication authentication) {
-        getAuthenticatedUser(authentication);
+        userService.getAuthenticatedUser(authentication);
         Pageable pageable = PageRequest.of(page,size);
         Page<CommunityPosts> communityPosts = communityPostRepository.findSanctionPosts(pageable);
 
@@ -50,7 +47,7 @@ public class CommunityManagementService {
     }
 
     public CommunityManagementPostDetailResponse getSanctionPostById(Long postId,Authentication authentication) {
-        getAuthenticatedUser(authentication);
+        userService.getAuthenticatedUser(authentication);
 
         CommunityPosts getSectionPost = communityPostRepository.findSanctionPostById(postId)
                 .orElseThrow(() -> new NotFoundException(ErrorMessage.COMMUNITY_POST_NOT_FOUND));
@@ -60,20 +57,11 @@ public class CommunityManagementService {
 
     @Transactional
     public void deleteSanctionPostById(Long postId, Authentication authentication) {
-        getAuthenticatedUser(authentication);
+        userService.getAuthenticatedUser(authentication);
 
         CommunityPosts getSectionPost = communityPostRepository.findById(postId)
                 .orElseThrow(() -> new NotFoundException(ErrorMessage.COMMUNITY_POST_NOT_FOUND));
         communityPostRepository.delete(getSectionPost);
-    }
-
-    public Long getAuthenticatedUserId(Authentication authentication) {
-        return ((CustomUserDetails) authentication.getPrincipal()).getUserId();
-    }
-
-    public User getAuthenticatedUser(Authentication authentication) {
-        return userRepository.findById(getAuthenticatedUserId(authentication))
-                .orElseThrow(() -> new UnauthorizedException(ErrorMessage.SESSION_EXPIRED));
     }
 
 }
